+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased][unreleased]
+#### Removed
+- the provided $DIR variable from header template
+
+#### Added
+- a set of support functions in the header template
 
 ### 0.1.1 - 2015-09-22
 #### Added